requirements.txt
allure-pytest==2.8.13
allure-python-commons==2.8.13
atomicwrites==1.3.0
attrs==19.3.0
importlib-metadata==1.6.0
more-itertools==8.2.0
packaging==20.3
pluggy==0.13.1
py==1.8.1
pyparsing==2.4.6
pytest==5.1.1
pytest-rerunfailures==7.0
selenium==3.13.0
six==1.14.0
wcwidth==0.1.9
zipp==3.1.0
